+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
-- Nothing.
+- Added the `IN` list comparison operator.
+- Added temporary `AmbiguousComponent` for cases where syntax isn't implemented by the library.
 
 ## [0.1.0] - 2018-01-25